+++ src/main/java/com/rematec/voucher/voucherbackapi/models/response/PromocaoResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -20,10 +21,10 @@ public class PromocaoResponse {
     private LocalDateTime fim;
     private String tipoDesconto;
     private String promocaoStatus;
-    private Double valorMinimoParaDisparo;
+    private BigDecimal valorMinimoParaDisparo;
     private Integer diasValidadeVoucher;
-    private Double discontoValor;
-    private Integer discontoPercentual;
+    private BigDecimal discontoValor;
+    private BigDecimal discontoPercentual;
     private String autorAlteracao;
 
     private List<LojaResponse> lojas;
